Read a texture's element type, and shift what a float register holds
dcl_resource_texture2d says whether a texture holds floats or integers, and
every texture but a Buffer or a multisampled one was written as the bare
Texture2D that means float4. A G-buffer packing a normal into the halves of a
uint then anded the texel as a float, which does not compile, and read the
depth beside it as whatever number those bits are. The declaration names the
type now, ld produces an integer where the resource holds one, and a float
reader of such a texel gets an asfloat rather than a conversion. The width is
left alone for a float texture: the component count comes from the reflection
data, which counts what the shader reads rather than what it declared.

A register whose components are all integers no float instruction touches is
declared int; one shared with a dot product is not, and holds its integer as
the float of the same value. HLSL refuses a shift of that outright, so the
operand is converted rather than reinterpreted. ushr had been converting with
(uint) alone, which is undefined on a negative.

Three fixtures from the sweep that found them. The instruction writer still
computes the wrong thing on the G-buffer shader, recorded as the fourth
instance of one wall: a mask of the low sixteen bits is a denormal as a float,
and fxc folds the asint of the asfloat of it to zero.

+++ HlslDecompiler.Tests/ShaderSources/cs_4_1_instruction/groupshared_reduce.fx
@@ -31,7 +31,7 @@ void main(CS_IN i)
 			g0[i.sv_groupindex.x] = r0.y;
 		}
 		GroupMemoryBarrierWithGroupSync();
-		r0.x = (uint)r0.x >> 1;
+		r0.x = (uint)(int)r0.x >> 1;
 	}
 	if (i.sv_groupindex.x == 0) {
 		r0.x = g0[0];
